IntDef: Update value type

+++ mobile/src/main/kotlin/info/papdt/express/helper/ui/fragment/settings/SettingsMain.kt
@@ -333,7 +333,7 @@ class SettingsMain : AbsPrefFragment(), Preference.OnPreferenceClickListener, Pr
 			}
 			// Query Api
 			mPrefApiType -> {
-				SettingsInstance.packageApiType = (o as String).toLong()
+				SettingsInstance.packageApiType = (o as String).toInt()
 				true
 			}
 			// Notification & push
